pyproject.toml
[tool.poetry]
name = "llmvm"
version = "0.1.0"
description = "Command Line LLM with client side tools support."
authors = ["Bilbo Baggins <contact@9600.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
async-timeout = "^4.0.3"
click = "^8.1.7"
jsonpickle = "^3.0.2"
nest-asyncio = "^1.6.0"
pyperclip = "^1.8.2"
rich = "^13.7.0"
openai = "^1.11.0"
anthropic = "^0.15.0"
mistralai = "^0.0.12"
click-default-group = "^1.2.4"
google-generativeai = "^0.3.2"
prompt-toolkit = "^3.0.43"
tiktoken = "^0.5.2"
psutil = "^5.9.8"
docstring-parser = "^0.15"
uvicorn = "^0.27.0.post1"
fastapi = "^0.109.0"
dill = "^0.3.8"
astunparse = "^1.6.3"
bs4 = "^0.0.2"
sec-api = "^1.0.17"
markdownify = "^0.11.6"
pillow = "^10.2.0"
aiofiles = "^23.2.1"
playwright = "^1.41.1"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
pdfminer = "^20191125"
pdfminer-six = "^20231228"
pypdf = "^4.0.1"
serpapi = "^0.1.5"
langchain = "^0.1.5"
python-multipart = "^0.0.6"
langchain-community = "^0.0.17"
python-dateutil = "^2.8.2"
googlesearch-python = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
